+++ src/main/java/com/gestaooserp/dev/repository/FabricanteRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.Fabricante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FabricanteRepository extends JpaRepository<Fabricante,Integer> {
+public interface FabricanteRepository extends JpaRepository<Fabricante,Long> {
 }
